Move model to db folder as suggested in previous PR
