+++ pkg/db/postgresql/model/validator_rewards.go
@@ -1,6 +1,6 @@
 package model
 
-import "github.com/cortze/eth2-state-analyzer/pkg/custom_spec"
+import "github.com/cortze/eth2-state-analyzer/pkg/fork_metrics/fork_state"
 
 // Postgres intregration variables
 var (
@@ -94,7 +94,7 @@ func NewValidatorRewards(
 		ValidatorIndex:       iValIdx,
 		Slot:                 int(iSlot),
 		Epoch:                int(iEpoch),
-		ValidatorBalance:     float32(iValBal) / float32(custom_spec.EFFECTIVE_BALANCE_INCREMENT),
+		ValidatorBalance:     float32(iValBal) / float32(fork_state.EFFECTIVE_BALANCE_INCREMENT),
 		Reward:               int(iReward),
 		MaxReward:            int(iMaxReward),
 		AttestationReward:    int(iMaxAttReward),
